medical_booking_app/requirements.txt
streamlit
google-cloud-bigquery
pandas
plotly
